refactor create node into its own file
